code/requirements.txt
cloudscraper
cssselect
fastapi
httpx
lxml
openai
pydantic
rembg
sseclient_py
uvicorn
Wand
wasmtime
